refactor(types): move led indicator to `rmk-types`, add
