+++ rmk-types/src/action.rs
@@ -19,6 +19,7 @@ impl Default for EncoderAction {
 }
 
 impl EncoderAction {
+    /// Create a new encoder action.
     pub const fn new(clockwise: KeyAction, counter_clockwise: KeyAction) -> Self {
         Self {
             clockwise,
@@ -26,18 +27,22 @@ impl EncoderAction {
         }
     }
 
+    /// Set the clockwise action.
     pub fn set_clockwise(&mut self, clockwise: KeyAction) {
         self.clockwise = clockwise;
     }
 
+    /// Set the counter clockwise action.
     pub fn set_counter_clockwise(&mut self, counter_clockwise: KeyAction) {
         self.counter_clockwise = counter_clockwise;
     }
 
+    /// Get the clockwise action.
     pub fn clockwise(&self) -> KeyAction {
         self.clockwise
     }
 
+    /// Get the counter clockwise action.
     pub fn counter_clockwise(&self) -> KeyAction {
         self.counter_clockwise
     }
