+++ rmk/src/event.rs
@@ -1,7 +1,7 @@
 use postcard::experimental::max_size::MaxSize;
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "controller")]
-use {crate::light::LedIndicator, rmk_types::action::KeyAction, rmk_types::modifier::ModifierCombination};
+use {rmk_types::action::KeyAction, rmk_types::led_indicator::LedIndicator, rmk_types::modifier::ModifierCombination};
 
 use crate::input_device::rotary_encoder::Direction;
