+++ rmk/src/storage/mod.rs
@@ -11,6 +11,7 @@ use embedded_storage::nor_flash::NorFlash;
 use embedded_storage_async::nor_flash::NorFlash as AsyncNorFlash;
 use heapless::Vec;
 use rmk_types::action::{EncoderAction, KeyAction};
+use rmk_types::led_indicator::LedIndicator;
 use rmk_types::modifier::ModifierCombination;
 use rmk_types::mouse_button::MouseButtons;
 use sequential_storage::Error as SSError;
@@ -27,7 +28,6 @@ use crate::channel::FLASH_CHANNEL;
 use crate::combo::Combo;
 use crate::config::{self, StorageConfig};
 use crate::fork::{Fork, StateBits};
-use crate::light::LedIndicator;
 use crate::morse::{Morse, MorseMode, MorsePattern};
 #[cfg(all(feature = "_ble", feature = "split"))]
 use crate::split::ble::PeerAddress;
